+++ CHANGELOG.md
@@ -29,7 +29,7 @@ Classify the change according to the following categories:
 ### Fixed
 - Align heating and cooling load profiles based on electric load year input, if using custom electric load profile with simulated (CRB or schedule-based flatloads) heating/cooling loads
 ### Changed
-- Make `year` input required with `ElectricLoad.loads_kw` input
+- Make `year` input required with any custom load profile input (e.g. `ElectricLoad.loads_kw`, `SpaceHeatingLoad.fuel_loads_mmbtu_per_hour`)
 - Shift and adjust CRB load profiles (i.e. with `doe_reference_name` input) based on the `year` input
 
 ## leap-year-fix