+++ CHANGELOG.md
@@ -37,6 +37,10 @@ Classify the change according to the following categories:
 - Handling of leap years for `ElectricLoad.loads_kw` inputs to align with URDB rate structures
 
 
+## v0.49.1
+### Changed
+- Swap an error for a warning with inconsistent load-year between electric and heating; soon to 
+
 ## v0.49.0
 ### Added
 - Ability to normalize and scale a custom load profile input to annual or monthly energy input values, for all load types